core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@2977 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/types/DateTimeType.java
@@ -1062,7 +1062,7 @@ public final class DateTimeType extends DTIType {
         return false;
     }
 
-    public long getPart(Session session, Object dateTime, int part) {
+    public int getPart(Session session, Object dateTime, int part) {
 
         int calendarPart;
         int increment = 0;
